fixup merge: move keyspace and models::partitioning into pageserver_api
