+++ pageserver/client/src/mgmt_api.rs
@@ -96,7 +96,7 @@ impl Client {
         &self,
         tenant_id: TenantId,
         timeline_id: TimelineId,
-    ) -> Result<crate::http::models::partitioning::Partitioning> {
+    ) -> Result<pageserver_api::models::partitioning::Partitioning> {
         let uri = format!(
             "{}/v1/tenant/{tenant_id}/timeline/{timeline_id}/keyspace",
             self.mgmt_api_endpoint
