--- pageserver/src/http/models.rs
@@ -1,3 +1,0 @@
-//! If possible, use `::pageserver_api::models` instead.
-
-pub mod partitioning;
